users namespace refactor
